update Bootstrap Recovery for the case of the virgin new PC without Tekla
On the new PC is TSmatch should setup User Environment Variable with the Path to TSmatch.xlsx.
It normally point to the Tekla Environment\exceldesign. However, when Tekla is not available, TSmatch.xlsx could be located anywhere.
When we manualy open this Excel file by double click in Windows, Bootstrap could pickup the Directory and set Environment Variable.